Some more refactoring / clean up
